ChatModels/1_chatModelOpenai.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model="gpt-4")

result = model.invoke("What is the capital of France?", temperature=0.7, max_completion_tokens=50)

print(result.content)